JBRULES-3200 Support for dynamic typing (aka "traits")
